requirements.txt
langchain_community==0.3.0
langchain_core==0.3.2
langchain_openai==0.2.0
langchain_text_splitters==0.3.0
python-dotenv==1.0.1
